Relativize trx and coverage filepaths to the workspace.
Some plugins (like "Archive Artifacts") use paths relative to the workspace.
This modification lets the user pass the results (trx, coverage) of this test run to other plugins.
This is a breaking change, but the user will still be able to concatenate the $WORKSPACE variable
to the variables set by vstestrunner-plugin.